REFACTOR: adjust API to a more logical structure
