+++ Cargo.toml
@@ -8,7 +8,7 @@ edition = "2021"
 exclude = ["/tmp"]
 homepage = "https://www.kpn.com/zakelijk/internet-of-things/data-services-hub.htm"
 keywords = ["kpn", "dsh", "sdk"]
-license = "Apache License 2.0" # TBD
+license = "Apache License 2.0"
 name = "dsh_sdk"
 readme = "README.md"
 repository = "https://github.com/kpn-dsh/dsh-sdk-platform-rs"
