+++ docker/docker-compose.yml
@@ -1,0 +1,82 @@
+version: "3"
+
+volumes:
+  postgres-data-volume: {}
+  pgadmin-volume: {}
+
+services:
+
+
+  zookeeper:
+    image: confluentinc/cp-zookeeper:latest
+    hostname: zookeeper
+    container_name: zookeeper
+    ports:
+      - "2181:2181"
+    environment:
+      ZOOKEEPER_CLIENT_PORT: 2181
+      ZOOKEEPER_TICK_TIME: 2000
+
+  kafka:
+    image: confluentinc/cp-kafka:latest
+    restart: always
+    hostname: kafka
+    container_name: kafka
+    depends_on:
+      - zookeeper
+    ports:
+      - "9092:9092"
+    environment:
+      KAFKA_BROKER_ID: 1
+      KAFKA_ZOOKEEPER_CONNECT: 'zookeeper:2181'
+      KAFKA_ADVERTISED_LISTENERS: PLAINTEXT://kafka:29092,PLAINTEXT_HOST://localhost:9092
+      KAFKA_LISTENER_SECURITY_PROTOCOL_MAP: PLAINTEXT:PLAINTEXT,PLAINTEXT_HOST:PLAINTEXT
+      KAFKA_INTER_BROKER_LISTENER_NAME: PLAINTEXT
+      KAFKA_OFFSETS_TOPIC_REPLICATION_FACTOR: 1
+      KAFKA_AUTO_CREATE_TOPICS_ENABLE: "true"
+
+  postgres:
+    image: postgres:latest
+    restart: always
+    container_name: postgres
+    environment:
+      - POSTGRES_USER=postgres
+      - POSTGRES_PASSWORD=postgres
+      - DATABASE_HOST=127.0.0.1
+    ports:
+      - '5432:5432'
+    volumes:
+      - postgres-data-volume:/var/lib/postgresql/data/
+
+  # Schema registry
+  apicurio:
+    image: apicurio/apicurio-registry-sql:2.4.4.Final
+    hostname: apicurio
+    container_name: apicurio
+    depends_on:
+      - postgres
+    restart: unless-stopped
+    ports:
+      - "8081:8080"
+    environment:
+      - REGISTRY_DATASOURCE_PASSWORD=postgres
+      - REGISTRY_DATASOURCE_USERNAME=postgres
+      - REGISTRY_DATASOURCE_URL=jdbc:postgresql://postgres:5432/postgres
+      - REGISTRY_RULES_GLOBAL_VALIDITY=SYNTAX_ONLY
+      - REGISTRY_RULES_GLOBAL_COMPATIBILITY=NONE
+      - REGISTRY_CCOMPAT_LEGACY_ID_MODE_ENABLED=true 
+      - REGISTRY_CCOMPAT_USE_CANONICAL_HASH=true
+      - REGISTRY_UI_FEATURES_READONLY=false
+      - REGISTRY_LIMITS_CONFIG_MAX_SCHEMA_SIZE_BYTES=5000
+      - REGISTRY_LIMITS_CONFIG_MAX_ARTIFACT_LABELS=0
+      - REGISTRY_LIMITS_CONFIG_MAX_ARTIFACT_PROPERTIES=0
+      - REGISTRY_LIMITS_CONFIG_MAX_ARTIFACTS=1000000000
+      - REGISTRY_LIMITS_CONFIG_MAX_DESCRIPTION_LENGTH=0
+      - REGISTRY_LIMITS_CONFIG_MAX_LABEL_SIZE=0
+      - REGISTRY_LIMITS_CONFIG_MAX_NAME_LENGTH=0
+      - REGISTRY_LIMITS_CONFIG_MAX_PROPERTY_KEY_SIZE=0
+      - REGISTRY_LIMITS_CONFIG_MAX_PROPERTY_VALUE_SIZE=0
+      - REGISTRY_LIMITS_CONFIG_MAX_VERSIONS_PER_ARTIFACT=5
+      - REGISTRY_LIMITS_CONFIG_MAX_TOTAL_SCHEMAS=1000000000
+      - REGISTRY_LIMITS_CONFIG_MAX_REQUESTS_PER_SECOND=300
+
