+++ src/error.rs
@@ -23,5 +23,7 @@ pub enum DshError {
     #[error("Error parsing Distinguished Name: {0}")]
     ParseDnError(String),
     #[error("Error getting group id, index out of bounds for {0}")]
-    IndexGroupIdError(crate::bootstrap::GroupType),
+    IndexGroupIdError(crate::kafka_properties::datastream::GroupType),
+    #[error("Error getting topic name {0}, Topic not found in datastreams." )]
+    NotFoundTopicError(String),
 }
